move assay.js to core, not under config
