+++ core/assay.js
@@ -1,10 +1,25 @@
 import harden from '@agoric/harden';
 
-import { insist } from '../../util/insist';
-import {
-  mustBeSameStructure,
-  mustBeComparable,
-} from '../../util/sameStructure';
+import { insist } from '../util/insist';
+import { mustBeSameStructure, mustBeComparable } from '../util/sameStructure';
+
+// This module treats labels as black boxes. It is not aware
+// of issuers, and so can handle labels whose issuers are merely
+// presences of remote issuers.
+
+// `makeAssayMaker` takes in a 'logic' object which defines set
+// operations for a particular kind of assay. For instance, for the
+// natAssay, these operations are arithmetic.
+
+// Return an assay, which makes amounts, validates amounts, and
+// provides set operations over amounts. An amount is a pass-by-copy
+// description of some set of erights. An amount has a label and a
+// quantity. All amounts made by the same assay have the same label
+// but differ in quantity.
+//
+// An assay is pass-by-presence, but is not designed to be usefully
+// passed. Rather, we expect each vat that needs to operate on amounts
+// will have its own local assay to do so.
 
 function makeAssayMaker(logic) {
   function makeAssay(label) {
