+++ more/pixels/pixelAssay.js
@@ -1,7 +1,7 @@
 import harden from '@agoric/harden';
 
 import { makePixelLogic } from './pixelLogic';
-import { makeAssayMaker } from '../../core/config/assay';
+import { makeAssayMaker } from '../../core/assay';
 
 // A pixelList is a naive collection of pixels in the form:
 // [ { x: 0, y: 0 }, { x: 1, y: 1} ...]
